add add and sub of RistrettoPoint and Ciphertext
Let $g$ be a base point, $x$ and $y$ be two public keys,
$m_1$ and $m_2$ be two ristretto points.
We can view a RistrettoPoint $(m_1)$ equivalent to Ciphertext $(m_1 + 0x, 0g)$,
and define the subtraction of RistrettoPoint and Ciphertext as
$(0g, m_1 + 0x) - (rg, m_2 + ry) = (rg, m_1 - m_2 + ry)$.
We can decipher the ciphertext with the secret key of the original
ciphertext.
This is can be useful in some case, for example,
we can calculate the subtraction of plaintext transaction fee
and ciphertext balance.
The resulting ciphertext can be deciphered with the original secret key.